Coloc 2: Encapsulate irregular IJ mask preparation in own method
Irregular IJ masks are the mask, contained in its bounding box. This
method will put this bounding box at the right place in a new image of the
same size as the original image (stack) of which the mask is part of.

This is done, because we expect the mask to be of the same size as the
image.