Coloc 2 - Require the offset and the size of a masks bounding box
This simplifies the use of masks, because a MaskLocalizableCursor can now
directly give the correct position relative to the bounding box. Of cause
other classes like DataContainer and MaskImageFactory have to know about
that, too.

The plugin class now distinguishes between irregular ROIs and masks. Inter-
nally they are treated the same, but will be constructed differently.
Currently only irregular ROIs are supported, because the BB information
is available right away.