Stack Focuser: add optional maximum projection feature
For the final image a pixel gets selected to be copied from a slice in the
original data. The change of this commit allows to look relative to the
selected pixel n slices above or below it if there is a pixel with a
higher value. If so, the higher value pixel is put into the final image.
The distance in which is looked (n) can be defined.