Color tool: mark the boundary structure before coloring
The color that is used to fill a cell might be the same as the
boundary color. If this happens the flood fill will break and
weird colors start to form. To circumvent this, the tool now
tries to find the boundary color, marking all the pixels containing
these values, doing the regular coloring and afterwards resetting
the boundary color to its original state.

The boundary color is found by using the first interior postion it
can get and walking to the left (or to the right if left is not
possible) until a different color is found. This is thes used as
the boundary color. This one is replaced with negative infinity
for every frame. The boundary color is reset after finishing the
colorization of the current frame.